Move FakeReactor to l.lib.testing and EventHandlingReactor to l.lib.reactor.
