+++ landscape/monitor/tests/test_service.py
@@ -1,7 +1,7 @@
 from mock import Mock
 
+from landscape.lib.testing import FakeReactor
 from landscape.tests.helpers import LandscapeTest, FakeBrokerServiceHelper
-from landscape.reactor import FakeReactor
 from landscape.monitor.config import MonitorConfiguration, ALL_PLUGINS
 from landscape.monitor.service import MonitorService
 from landscape.monitor.computerinfo import ComputerInfo
