+++ landscape/package/tests/test_changer.py
@@ -11,7 +11,7 @@ from twisted.internet.error import ProcessTerminated, ProcessDone
 from mock import patch, Mock, call
 
 from landscape.lib.fs import create_text_file, read_text_file, touch_file
-from landscape.lib.testing import StubProcessFactory
+from landscape.lib.testing import StubProcessFactory, FakeReactor
 from landscape.package.changer import (
     PackageChanger, main, UNKNOWN_PACKAGE_DATA_TIMEOUT,
     SUCCESS_RESULT, DEPENDENCY_ERROR_RESULT, POLICY_ALLOW_INSTALLS,
@@ -27,7 +27,6 @@ from landscape.package.tests.helpers import (
     AptFacadeHelper, SimpleRepositoryHelper)
 from landscape.manager.manager import FAILED
 from landscape.manager.shutdownmanager import ShutdownFailedError
-from landscape.reactor import FakeReactor
 
 
 class AptPackageChangerTest(LandscapeTest):
