+++ landscape/tests/test_amp.py
@@ -2,10 +2,10 @@ from twisted.internet.error import ConnectError
 from twisted.internet.task import Clock
 
 from landscape.tests.helpers import LandscapeTest
-from landscape.reactor import FakeReactor
 from landscape.deployment import Configuration
 from landscape.amp import ComponentPublisher, ComponentConnector, remote
 from landscape.lib.amp import MethodCallError
+from landscape.lib.testing import FakeReactor
 
 
 class TestComponent(object):
